+++ frontend/src/components/Footer.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-const VERSION = "0.3.1";
+const VERSION = "0.4.0";
 
 class Footer extends Component {
     render() {